AT-SRC-001 dtype neutrality fix — Cluster C3 RESOLVED tests/ passed 516, failed 27, skipped 143
[SOURCE-WEIGHT-002] Phase D2 complete — Attempt #19

Changes:
- tests/test_at_src_001_simple.py:83,93: derive dtype from parser output
  instead of hardcoding float32

Test results:
- Targeted: 10/10 PASSED (100%)
- Full suite: 516 passed, 27 failed (was 31), 143 skipped, 2 xfailed
- Runtime: 1828.32s (30m 28s)
- Net improvement: -4 failures (31→27, -12.9%)

Cluster C3 status:
- Phase K baseline: 4 failures
- Phase D result: 0 failures (✅ RESOLVED)
- Coverage: 33/33 source weighting tests PASSED

Exit criteria met:
✅ All AT-SRC-001 tests pass
✅ dtype propagation validated (float32/float64)
✅ Default matches torch.get_default_dtype()
✅ Full suite -4 failures (C3 cleared)

Artifacts: reports/2026-01-test-suite-triage/phase_d/20251011T093344Z/
Status: [SOURCE-WEIGHT-002] marked done

+++ tests/test_at_src_001_simple.py
@@ -80,7 +80,8 @@ def test_sourcefile_parsing():
 
         # Check directions are normalized
         norms = torch.linalg.norm(directions, dim=1)
-        torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float32))
+        # Use dtype from parser output (defaults to torch.get_default_dtype())
+        torch.testing.assert_close(norms, torch.ones(2, dtype=directions.dtype))
 
         # Per spec-a-core.md:151-153, wavelength column is IGNORED.
         # All sources use CLI -lambda value (default_wavelength_m).
@@ -88,7 +89,8 @@ def test_sourcefile_parsing():
         assert wavelengths[1].item() == pytest.approx(default_wavelength_m)
 
         # Check weights are preserved from file (for parsing correctness)
-        expected_weights = torch.tensor([2.0, 3.0], dtype=torch.float32)
+        # Use dtype from parser output (defaults to torch.get_default_dtype())
+        expected_weights = torch.tensor([2.0, 3.0], dtype=weights.dtype)
         torch.testing.assert_close(weights, expected_weights)
 
         print(f"✓ Source file parsing successful")
